# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  ad.StoreRequest:
    type: object
  ad.UpdateRequest:
    type: object
...
  chat.ChatResponse:
    properties:
      ad_id:
        type: integer
      buyer_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      seller_id:
        type: integer
    type: object
  chat.ListChatResponse:
    properties:
      chats:
        items:
          $ref: '#/definitions/chat.ChatResponse'
        type: array
    type: object
  chat.ListMessageResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/chat.MessageResponse'
        type: array
    type: object
  chat.MessageResponse:
    properties:
      chat_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_read:
        type: boolean
      mine:
        type: boolean
      sender_id:
        type: integer
      text:
        type: string
    type: object
  chat.StoreChatRequest:
    properties:
      ad_id:
        type: integer
      seller_id:
        type: integer
    type: object
  chat.StoreMessageRequest:
    properties:
      chat_id:
        type: integer
      text:
        type: string
    type: object
  handler.BasicResponseDTO:
    properties:
      data: {}
      status:
        type: integer
    type: object
  handler.ErrorResponse:
    properties:
      error:
        example: chat
        type: string
    type: object
...
host: http://localhost:8989
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Ads Service API
  version: "1.0"
paths:
...
  /api/v1/ads/{id}:
    delete:
      consumes:
      - application/json
      description: Delete car advertisement
      parameters:
      - description: Ad data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ad.StoreRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.BasicResponseDTO'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete advertisement
      tags:
      - Ads
    get:
      consumes:
      - application/json
      description: Get car advertisement details
      parameters:
      - description: Ad ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.BasicResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get advertisement by ID
      tags:
      - Ads
    patch:
      consumes:
      - application/json
      description: Update car advertisement
      parameters:
      - description: Ad data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ad.UpdateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.BasicResponseDTO'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update advertisement
      tags:
      - Ads
...
  /api/v1/chats:
    get:
      consumes:
      - application/json
      description: Get chats list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.BasicResponseDTO'
            - properties:
                data:
                  $ref: '#/definitions/chat.ListChatResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get chats list
      tags:
      - Chats
    post:
      consumes:
      - application/json
      description: Create new chat
      parameters:
      - description: Chat data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/chat.StoreChatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.BasicResponseDTO'
            - properties:
                data:
                  $ref: '#/definitions/chat.ChatResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create new chat
      tags:
      - Chats
  /api/v1/chats/{id}/message:
    post:
      consumes:
      - application/json
      description: Create new message
      parameters:
      - description: Message data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/chat.StoreMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.BasicResponseDTO'
            - properties:
                data:
                  $ref: '#/definitions/chat.MessageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create new message
      tags:
      - Chats
  /api/v1/chats/{id}/messages:
    get:
      consumes:
      - application/json
      description: Get messages list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.BasicResponseDTO'
            - properties:
                data:
                  $ref: '#/definitions/chat.ListMessageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get messages list
      tags:
      - Chats
...
swagger: "2.0"